+++ src/modules/NetworkLogger.ts
@@ -2,15 +2,32 @@ import type { RequestHandler } from '@apollo/client';
 
 import InstabugConstants from '../utils/InstabugConstants';
 import xhr, { NetworkData, ProgressCallback } from '../utils/XhrNetworkInterceptor';
-import { isContentTypeNotAllowed, reportNetworkLog } from '../utils/InstabugUtils';
 import { InstabugRNConfig } from '../utils/config';
 import { Logger } from '../utils/logger';
+import { NativeInstabug } from '../native/NativeInstabug';
+import {
+  isContentTypeNotAllowed,
+  registerFilteringAndObfuscationListener,
+  registerFilteringListener,
+  registerObfuscationListener,
+  reportNetworkLog,
+} from '../utils/InstabugUtils';
+import {
+  NativeNetworkLogger,
+  NativeNetworkLoggerEvent,
+  NetworkListenerType,
+  NetworkLoggerEmitter,
+} from '../native/NativeNetworkLogger';
+import { Platform } from 'react-native';
 
 export type { NetworkData };
 
 export type NetworkDataObfuscationHandler = (data: NetworkData) => Promise<NetworkData>;
 let _networkDataObfuscationHandler: NetworkDataObfuscationHandler | null | undefined;
 let _requestFilterExpression = 'false';
+let _isNativeInterceptionEnabled = false;
+let _networkListener: NetworkListenerType | null = null;
+let hasFilterExpression = false;
 
 function getPortFromUrl(url: string) {
   const portMatch = url.match(/:(\d+)(?=\/|$)/);
@@ -76,6 +93,22 @@ export const setEnabled = (isEnabled: boolean) => {
 };
 
 /**
+ * @internal
+ * Sets whether enabling or disabling native network interception.
+ * It is disabled by default.
+ * @param isEnabled
+ */
+export const setNativeInterceptionEnabled = (isEnabled: boolean) => {
+  _isNativeInterceptionEnabled = isEnabled;
+};
+
+export const getNetworkDataObfuscationHandler = () => _networkDataObfuscationHandler;
+
+export const getRequestFilterExpression = () => _requestFilterExpression;
+
+export const hasRequestFilterExpression = () => hasFilterExpression;
+
+/**
  * Obfuscates any response data.
  * @param handler
  */
@@ -83,6 +116,13 @@ export const setNetworkDataObfuscationHandler = (
   handler?: NetworkDataObfuscationHandler | null | undefined,
 ) => {
   _networkDataObfuscationHandler = handler;
+  if (_isNativeInterceptionEnabled && Platform.OS === 'ios') {
+    if (hasFilterExpression) {
+      registerFilteringAndObfuscationListener(_requestFilterExpression);
+    } else {
+      registerObfuscationListener();
+    }
+  }
 };
 
 /**
@@ -91,6 +131,15 @@ export const setNetworkDataObfuscationHandler = (
  */
 export const setRequestFilterExpression = (expression: string) => {
   _requestFilterExpression = expression;
+  hasFilterExpression = true;
+
+  if (_isNativeInterceptionEnabled && Platform.OS === 'ios') {
+    if (_networkDataObfuscationHandler) {
+      registerFilteringAndObfuscationListener(_requestFilterExpression);
+    } else {
+      registerFilteringListener(_requestFilterExpression);
+    }
+  }
 };
 
 /**
@@ -113,4 +162,93 @@ export const apolloLinkRequestHandler: RequestHandler = (operation, forward) => 
   }
 
   return forward(operation);
+};
+
+/**
+ * Sets whether network body logs will be captured or not.
+ * @param isEnabled
+ */
+export const setNetworkLogBodyEnabled = (isEnabled: boolean) => {
+  NativeInstabug.setNetworkLogBodyEnabled(isEnabled);
+};
+
+/**
+ * @internal
+ * Exported for internal/testing purposes only.
+ */
+export const resetNetworkListener = () => {
+  if (process.env.NODE_ENV === 'test') {
+    _networkListener = null;
+    NativeNetworkLogger.resetNetworkLogsListener();
+  } else {
+    Logger.error(
+      `${InstabugConstants.IBG_APM_TAG}: The \`resetNetworkListener()\` method is intended solely for testing purposes.`,
+    );
+  }
+};
+
+/**
+ * @internal
+ * Exported for internal/testing purposes only.
+ */
+export const registerNetworkLogsListener = (
+  type: NetworkListenerType,
+  handler?: (networkSnapshot: NetworkData) => void,
+) => {
+  if (Platform.OS === 'ios') {
+    // remove old listeners
+    if (NetworkLoggerEmitter.listenerCount(NativeNetworkLoggerEvent.NETWORK_LOGGER_HANDLER) > 0) {
+      NetworkLoggerEmitter.removeAllListeners(NativeNetworkLoggerEvent.NETWORK_LOGGER_HANDLER);
+    }
+
+    if (_networkListener == null) {
+      // set new listener.
+      _networkListener = type;
+    } else {
+      // attach a new listener to the existing one.
+      _networkListener = NetworkListenerType.both;
+    }
+  }
+
+  NetworkLoggerEmitter.addListener(
+    NativeNetworkLoggerEvent.NETWORK_LOGGER_HANDLER,
+    (networkSnapshot) => {
+      // Mapping the data [Native -> React-Native].
+      const { id, url, requestHeader, requestBody, responseHeader, response, responseCode } =
+        networkSnapshot;
+
+      const networkSnapshotObj: NetworkData = {
+        id: id,
+        url: url,
+        requestBody: requestBody,
+        requestHeaders: requestHeader,
+        method: '',
+        responseBody: response,
+        responseCode: responseCode,
+        responseHeaders: responseHeader,
+        contentType: '',
+        duration: 0,
+        requestBodySize: 0,
+        responseBodySize: 0,
+        errorDomain: '',
+        errorCode: 0,
+        startTime: 0,
+        serverErrorMessage: '',
+        requestContentType: '',
+        isW3cHeaderFound: true,
+        networkStartTimeInSeconds: 0,
+        partialId: 0,
+        w3cCaughtHeader: '',
+        w3cGeneratedHeader: '',
+      };
+      if (handler) {
+        handler(networkSnapshotObj);
+      }
+    },
+  );
+  if (Platform.OS === 'ios') {
+    NativeNetworkLogger.registerNetworkLogsListener(_networkListener ?? NetworkListenerType.both);
+  } else {
+    NativeNetworkLogger.registerNetworkLogsListener();
+  }
 };