+++ src/native/NativeInstabug.ts
@@ -2,6 +2,7 @@ import { NativeEventEmitter, NativeModule, ProcessedColorValue } from 'react-nat
 
 import type Report from '../models/Report';
 import type {
+  AutoMaskingType,
   ColorTheme,
   InvocationEvent,
   Locale,
@@ -72,6 +73,7 @@ export interface InstabugNativeModule extends NativeModule {
   ): void;
 
   setNetworkLoggingEnabled(isEnabled: boolean): void;
+  setNetworkLogBodyEnabled(isEnabled: boolean): void;
 
   // Repro Steps APIs //
   setReproStepsConfig(
@@ -152,12 +154,16 @@ export interface InstabugNativeModule extends NativeModule {
 
   // W3C Feature Flags Listener for Android
   registerW3CFlagsChangeListener(): void;
+
+  setOnFeaturesUpdatedListener(handler?: (params: any) => void): void; // android only
+  enableAutoMasking(autoMaskingTypes: AutoMaskingType[]): void;
 }
 
 export const NativeInstabug = NativeModules.Instabug;
 
 export enum NativeEvents {
   PRESENDING_HANDLER = 'IBGpreSendingHandler',
+  IBG_ON_FEATURES_UPDATED_CALLBACK = 'IBGOnFeatureUpdatedCallback',
   ON_W3C_FLAGS_CHANGE = 'IBGOnNewW3CFlagsUpdateReceivedCallback',
 }