+++ test/mocks/mockInstabug.ts
@@ -73,6 +73,9 @@ const mockInstabug: InstabugNativeModule = {
   isW3ExternalGeneratedHeaderEnabled: jest.fn(),
   isW3CaughtHeaderEnabled: jest.fn(),
   registerW3CFlagsChangeListener: jest.fn(),
+  setNetworkLogBodyEnabled: jest.fn(),
+  setOnFeaturesUpdatedListener: jest.fn(),
+  enableAutoMasking: jest.fn(),
 };
 
 export default mockInstabug;