+++ test/modules/NetworkLogger.spec.ts
@@ -7,14 +7,27 @@ import * as NetworkLogger from '../../src/modules/NetworkLogger';
 import Interceptor from '../../src/utils/XhrNetworkInterceptor';
 import { isContentTypeNotAllowed, reportNetworkLog } from '../../src/utils/InstabugUtils';
 import InstabugConstants from '../../src/utils/InstabugConstants';
+import * as Instabug from '../../src/modules/Instabug';
+import {
+  NativeNetworkLogger,
+  NativeNetworkLoggerEvent,
+  NetworkListenerType,
+  NetworkLoggerEmitter,
+} from '../../src/native/NativeNetworkLogger';
+import { InvocationEvent, LogLevel, NetworkInterceptionMode } from '../../src';
+import { Platform } from 'react-native';
 import { Logger } from '../../src/utils/logger';
+import { NativeInstabug } from '../../src/native/NativeInstabug';
 
 const clone = <T>(obj: T): T => {
   return JSON.parse(JSON.stringify(obj));
 };
 
+jest.mock('../../src/native/NativeNetworkLogger');
+
 describe('NetworkLogger Module', () => {
   const network: NetworkLogger.NetworkData = {
+    id: '',
     url: 'https://api.instabug.com',
     requestBody: '',
     requestHeaders: { 'content-type': 'application/json' },
@@ -281,5 +294,126 @@ describe('NetworkLogger Module', () => {
     NetworkLogger.setEnabled(true);
 
     expect(reportNetworkLog).toHaveBeenCalledWith(networkData);
+  });
+
+  it('should call the native method setNetworkLogBodyEnabled', () => {
+    NetworkLogger.setNetworkLogBodyEnabled(true);
+
+    expect(NativeInstabug.setNetworkLogBodyEnabled).toBeCalledTimes(1);
+    expect(NativeInstabug.setNetworkLogBodyEnabled).toBeCalledWith(true);
+  });
+
+  it('Instabug.init should call NativeNetworkLogger.isNativeInterceptionEnabled and not call NativeNetworkLogger.hasAPMNetworkPlugin with iOS', async () => {
+    Platform.OS = 'ios';
+    const config = {
+      token: 'some-token',
+      invocationEvents: [InvocationEvent.floatingButton, InvocationEvent.shake],
+      debugLogsLevel: LogLevel.debug,
+      networkInterceptionMode: NetworkInterceptionMode.native,
+      codePushVersion: '1.1.0',
+    };
+    await Instabug.init(config);
+
+    expect(NativeNetworkLogger.isNativeInterceptionEnabled).toHaveBeenCalled();
+    expect(NativeNetworkLogger.hasAPMNetworkPlugin).not.toHaveBeenCalled();
+  });
+});
+
+describe('_registerNetworkLogsListener', () => {
+  let handlerMock: jest.Mock;
+  let type: NetworkListenerType;
+
+  beforeEach(() => {
+    handlerMock = jest.fn();
+    type = NetworkListenerType.both;
+    jest.resetAllMocks(); // Reset mock implementation and calls
+    NetworkLogger.resetNetworkListener(); // Clear only calls, keeping implementation intact
+  });
+
+  it('should remove old listeners if they exist', () => {
+    Platform.OS = 'ios';
+
+    // Simulate that there are existing listeners
+    jest.spyOn(NetworkLoggerEmitter, 'listenerCount').mockReturnValue(2);
+
+    NetworkLogger.registerNetworkLogsListener(type, handlerMock);
+
+    expect(NetworkLoggerEmitter.removeAllListeners).toHaveBeenCalledWith(
+      NativeNetworkLoggerEvent.NETWORK_LOGGER_HANDLER,
+    );
+  });
+
+  it('should set the new listener if _networkListener is null', () => {
+    Platform.OS = 'ios';
+    // No existing listener
+    jest.spyOn(NetworkLoggerEmitter, 'listenerCount').mockReturnValue(0);
+
+    NetworkLogger.registerNetworkLogsListener(type, handlerMock);
+
+    expect(NetworkLoggerEmitter.addListener).toHaveBeenCalled();
+    expect(NativeNetworkLogger.registerNetworkLogsListener).toHaveBeenCalledWith(type);
+  });
+
+  it('should attach a new listener to the existing one if _networkListener is set', () => {
+    Platform.OS = 'ios';
+
+    type = NetworkListenerType.filtering;
+    const newType = NetworkListenerType.both;
+
+    // First call to set the listener
+    NetworkLogger.registerNetworkLogsListener(type, handlerMock);
+
+    // Second call with a different type to trigger setting to `both`
+    NetworkLogger.registerNetworkLogsListener(newType, handlerMock);
+
+    expect(NetworkLoggerEmitter.addListener).toHaveBeenCalledTimes(2);
+    expect(NativeNetworkLogger.registerNetworkLogsListener).toHaveBeenCalledWith(
+      NetworkListenerType.both,
+    );
+  });
+
+  it('should map networkSnapshot data correctly and call handler', () => {
+    const mockNetworkSnapshot = {
+      id: '123',
+      url: 'http://example.com',
+      requestHeader: {},
+      requestBody: 'test request body',
+      responseHeader: {},
+      response: 'test response',
+      responseCode: 200,
+    };
+
+    (NetworkLoggerEmitter.addListener as jest.Mock).mockImplementation((_, callback) => {
+      callback(mockNetworkSnapshot);
+    });
+
+    NetworkLogger.registerNetworkLogsListener(type, handlerMock);
+
+    const expectedNetworkData: NetworkLogger.NetworkData = {
+      id: '123',
+      url: 'http://example.com',
+      requestBody: 'test request body',
+      requestHeaders: {},
+      method: '',
+      responseBody: 'test response',
+      responseCode: 200,
+      responseHeaders: {},
+      contentType: '',
+      duration: 0,
+      requestBodySize: 0,
+      responseBodySize: 0,
+      errorDomain: '',
+      errorCode: 0,
+      startTime: 0,
+      serverErrorMessage: '',
+      requestContentType: '',
+      isW3cHeaderFound: true,
+      networkStartTimeInSeconds: 0,
+      partialId: 0,
+      w3cCaughtHeader: '',
+      w3cGeneratedHeader: '',
+    };
+
+    expect(handlerMock).toHaveBeenCalledWith(expectedNetworkData);
   });
 });